Projects/PyCharm/Scripts/PyTorchAlexNetPerformance/requirements.txt
torch
Pillow
torchvision